requirements.txt
# python==3.10.16
PyPDF2
langchain
langchain_community
langchain_core
langchain_chroma
langchain_text_splitters
google-generativeai
python-dotenv
ipykernel
streamlit
mysql-connector-python
SQLAlchemy
langchain_groq
sentence-transformers
faiss-cpu
InstructorEmbedding